Small fix for posix platforms.
Actually sys/unistd.h is often not present, and when it's not,
you also get errors with sys/signal.h. Tgis uses header checl
to see if sys/unistd.h exists and falls back to <unistd.h>.
Depending on what is done here, it might be safer to use
<csignal> ubstead. Or maybe <unistd.h> / <signal.h> should be the
default.  Hard to say since I may not test on the same platforms
you have, so the header check seemed safest.